% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Criteria.R
\docType{class}
\name{Criteria}
\alias{Criteria}
\title{Criteria Reference Class}
\description{
A class to represent criteria with an operation, field name, and field values.
}
\section{Fields}{

\describe{
\item{\code{operation}}{A character string representing the operation.}

\item{\code{field_name}}{A character string representing the field name.}

\item{\code{field_values}}{A list of values for the field.}
}}


