% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{Query}
\alias{Query}
\title{Query Reference Class}
\description{
A class to represent a query with AND, OR, and NOT criteria.
}
\section{Fields}{

\describe{
\item{\code{and_criteria}}{A list of Criteria objects for AND conditions.}

\item{\code{or_criteria}}{A list of Criteria objects for OR conditions.}

\item{\code{not_criteria}}{A list of Criteria objects for NOT conditions.}
}}


