% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryUserFollowingRequest.R
\docType{class}
\name{QueryUserFollowingRequest}
\alias{QueryUserFollowingRequest}
\title{QueryUserFollowingRequest Reference Class}
\description{
A class to represent a request for the list of users that a specific user is following.
}
\section{Fields}{

\describe{
\item{\code{username}}{A character string representing the username.}

\item{\code{max_count}}{An optional numeric value representing the maximum count of users to retrieve.}

\item{\code{cursor}}{An optional character string for the cursor.}
}}


