% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryUserInfoRequest.R
\docType{class}
\name{QueryUserInfoRequest}
\alias{QueryUserInfoRequest}
\title{QueryUserInfoRequest Reference Class}
\description{
A class to represent a user info request with various parameters.
}
\section{Fields}{

\describe{
\item{\code{username}}{A character string representing the username.}

\item{\code{fields}}{An optional character string of fields to be retrieved.}
}}


