% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryUserLikedVideosRequest.R
\docType{class}
\name{QueryUserLikedVideosRequest}
\alias{QueryUserLikedVideosRequest}
\title{QueryUserLikedVideosRequest Reference Class}
\description{
A class to represent a request for user liked videos with various parameters.
}
\section{Fields}{

\describe{
\item{\code{username}}{A character string representing the username.}

\item{\code{max_count}}{An optional  numeric value representing the maximum count of videos to retrieve.}

\item{\code{cursor}}{An optional character string for the cursor.}

\item{\code{fields}}{An optional character string representing the fields to retrieve.}
}}


