% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryUserPinnedVideos.R
\docType{class}
\name{QueryUserPinnedVideosRequest}
\alias{QueryUserPinnedVideosRequest}
\title{QueryUserPinnedVideosRequest Reference Class}
\description{
A class to represent a request for user pinned videos with various parameters.
}
\section{Fields}{

\describe{
\item{\code{username}}{A character string representing the username.}

\item{\code{fields}}{An optional character string representing the fields to retrieve.}
}}


