% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryUserRespostedVideos.R
\docType{class}
\name{QueryUserRepostedVideosRequest}
\alias{QueryUserRepostedVideosRequest}
\title{QueryUserRepostedVideosRequest Reference Class}
\description{
A class to represent a request for user reposted videos with various parameters.
}
\section{Fields}{

\describe{
\item{\code{username}}{A character string representing the username.}

\item{\code{fields}}{An optional character string representing the fields to retrieve.}

\item{\code{cursor}}{An optional character string for the cursor.}

\item{\code{max_count}}{An optional numeric value representing the maximum count of videos to retrieve.}
}}


