% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryVideoCommentsRequest.R
\docType{class}
\name{QueryVideoCommentsRequest}
\alias{QueryVideoCommentsRequest}
\title{QueryVideoCommentsRequest Reference Class}
\description{
A class to represent a request for video comments with various parameters.
}
\section{Fields}{

\describe{
\item{\code{video_id}}{A character string representing the video ID.}

\item{\code{max_count}}{An optional  numeric value representing the maximum count of comments to retrieve.}

\item{\code{cursor}}{An optional character string for the cursor.}

\item{\code{fields}}{An optional character string representing the fields to retrieve.}
}}


