% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryVideoRequest.R
\docType{class}
\name{QueryVideoRequest}
\alias{QueryVideoRequest}
\title{QueryVideoRequest Reference Class}
\description{
A class to represent a query video request with various parameters.
}
\section{Fields}{

\describe{
\item{\code{query}}{A Query object representing the query.}

\item{\code{start_date}}{A character string representing the start date.}

\item{\code{end_date}}{A character string representing the end date.}

\item{\code{max_count}}{An optional  numeric value representing the maximum count.}

\item{\code{cursor}}{An optional character string for the cursor.}

\item{\code{is_random}}{An optional character value indicating if the request is random.}

\item{\code{search_id}}{An optional character string for the search ID.}

\item{\code{fields}}{An optional character string representing the fields to retrieve.}
}}


