% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TikTokResearchAPI.R
\docType{class}
\name{TikTokResearchAPI-class}
\alias{TikTokResearchAPI-class}
\alias{TikTokResearchAPI}
\title{TikTok Research API Client}
\description{
A class to interact with the TikTok Research API.
}

